refactor(auth): extract test helpers to shared modules
