+++ packages/auth/tests/integration/interceptor-cache-propagation.test.ts
@@ -15,26 +15,7 @@ import { parseAuthHeaders, setAuthHeaders } from "../../src/headers.ts";
 import { createSessionAuthInterceptor } from "../../src/session-auth-interceptor.ts";
 import type { AuthContext } from "../../src/types.ts";
 import { AUTH_HEADERS } from "../../src/types.ts";
-
-/** Create a mock ConnectRPC request for testing interceptors. */
-function createMockRequest(options?: {
-    serviceName?: string;
-    methodName?: string;
-    headers?: Headers;
-}) {
-    const serviceName = options?.serviceName ?? "test.v1.TestService";
-    const methodName = options?.methodName ?? "TestMethod";
-    const headers = options?.headers ?? new Headers();
-
-    return {
-        service: { typeName: serviceName },
-        method: { name: methodName },
-        header: headers,
-        url: `http://localhost/${serviceName}/${methodName}`,
-        stream: false,
-        message: {},
-    } as any;
-}
+import { createMockRequest } from "../helpers/mock-request.ts";
 
 describe("Interceptor Cache & Propagation — Integration", () => {
     describe("cache via createAuthInterceptor", () => {
