+++ packages/auth/tests/unit/proto-reader.test.ts
@@ -7,95 +7,14 @@
 
 import assert from "node:assert";
 import { describe, it } from "node:test";
-import type { DescMethod, DescService } from "@bufbuild/protobuf";
-import { create, setExtension } from "@bufbuild/protobuf";
-import { MethodOptionsSchema, ServiceOptionsSchema } from "@bufbuild/protobuf/wkt";
-import { AuthRequirementsSchema, MethodAuthSchema, method_auth, ServiceAuthSchema, service_auth } from "#gen/connectum/auth/v1/options_pb.js";
 import { getPublicMethods, resolveMethodAuth } from "../../src/proto/reader.ts";
-
-/**
- * Create a fake DescService with optional proto service options.
- */
-function createFakeService(options?: { typeName?: string; serviceOptions?: unknown }): DescService {
-    return {
-        kind: "service",
-        typeName: options?.typeName ?? "test.v1.TestService",
-        name: "TestService",
-        methods: [],
-        method: {},
-        deprecated: false,
-        proto: { options: options?.serviceOptions },
-    } as unknown as DescService;
-}
-
-/**
- * Create a fake DescMethod attached to a service, with optional proto method options.
- */
-function createFakeMethod(service: DescService, name: string, methodOptions?: unknown): DescMethod {
-    const method = {
-        kind: "rpc",
-        name,
-        localName: name.charAt(0).toLowerCase() + name.slice(1),
-        parent: service,
-        methodKind: "unary",
-        deprecated: false,
-        proto: { options: methodOptions },
-    } as unknown as DescMethod;
-    (service.methods as DescMethod[]).push(method);
-    return method;
-}
-
-/**
- * Create MethodOptions with method_auth extension set.
- */
-function createMethodOptions(authConfig: { public?: boolean; policy?: string; requires?: { roles?: string[]; scopes?: string[] } }) {
-    const opts = create(MethodOptionsSchema);
-    const init: Record<string, unknown> = {
-        policy: authConfig.policy ?? "",
-    };
-    // Only set public when explicitly provided to preserve proto2 field presence semantics
-    if (authConfig.public !== undefined) {
-        init.public = authConfig.public;
-    }
-    if (authConfig.requires) {
-        init.requires = create(AuthRequirementsSchema, {
-            roles: authConfig.requires.roles ?? [],
-            scopes: authConfig.requires.scopes ?? [],
-        });
-    }
-    const authMsg = create(MethodAuthSchema, init as any);
-    setExtension(opts, method_auth, authMsg);
-    return opts;
-}
-
-/**
- * Create ServiceOptions with service_auth extension set.
- */
-function createServiceOptions(authConfig: { defaultPolicy?: string; public?: boolean; defaultRequires?: { roles?: string[]; scopes?: string[] } }) {
-    const opts = create(ServiceOptionsSchema);
-    const init: Record<string, unknown> = {
-        defaultPolicy: authConfig.defaultPolicy ?? "",
-    };
-    // Only set public when explicitly provided to preserve proto2 field presence semantics
-    if (authConfig.public !== undefined) {
-        init.public = authConfig.public;
-    }
-    if (authConfig.defaultRequires) {
-        init.defaultRequires = create(AuthRequirementsSchema, {
-            roles: authConfig.defaultRequires.roles ?? [],
-            scopes: authConfig.defaultRequires.scopes ?? [],
-        });
-    }
-    const authMsg = create(ServiceAuthSchema, init as any);
-    setExtension(opts, service_auth, authMsg);
-    return opts;
-}
+import { createFakeMethod, createFakeService, createMethodOptions, createServiceOptions } from "../helpers/proto-test-helpers.ts";
 
 describe("proto-reader", () => {
     describe("resolveMethodAuth()", () => {
         it("should return defaults when no proto options are set", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "PlainMethod");
+            const method = createFakeMethod(service, "PlainMethod", undefined, { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -106,7 +25,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with public: true", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "PublicMethod", createMethodOptions({ public: true }));
+            const method = createFakeMethod(service, "PublicMethod", createMethodOptions({ public: true }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -116,7 +35,7 @@ describe("proto-reader", () => {
         it("should inherit service-level public: true", () => {
             const svcOpts = createServiceOptions({ public: true });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "AnyMethod");
+            const method = createFakeMethod(service, "AnyMethod", undefined, { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -125,7 +44,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with required roles", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "AdminOnly", createMethodOptions({ requires: { roles: ["admin"] } }));
+            const method = createFakeMethod(service, "AdminOnly", createMethodOptions({ requires: { roles: ["admin"] } }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -137,7 +56,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with required scopes", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "ScopedMethod", createMethodOptions({ requires: { scopes: ["read", "write"] } }));
+            const method = createFakeMethod(service, "ScopedMethod", createMethodOptions({ requires: { scopes: ["read", "write"] } }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -148,7 +67,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with roles and scopes combined", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "RolesAndScopes", createMethodOptions({ requires: { roles: ["user"], scopes: ["read"] } }));
+            const method = createFakeMethod(service, "RolesAndScopes", createMethodOptions({ requires: { roles: ["user"], scopes: ["read"] } }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -159,7 +78,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with policy: allow", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "AllowMethod", createMethodOptions({ policy: "allow" }));
+            const method = createFakeMethod(service, "AllowMethod", createMethodOptions({ policy: "allow" }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -168,7 +87,7 @@ describe("proto-reader", () => {
 
         it("should resolve method with policy: deny", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "DenyMethod", createMethodOptions({ policy: "deny" }));
+            const method = createFakeMethod(service, "DenyMethod", createMethodOptions({ policy: "deny" }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -178,7 +97,7 @@ describe("proto-reader", () => {
         it("should inherit service default_policy when method has no policy", () => {
             const svcOpts = createServiceOptions({ defaultPolicy: "deny" });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "InheritsPolicy");
+            const method = createFakeMethod(service, "InheritsPolicy", undefined, { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -188,7 +107,7 @@ describe("proto-reader", () => {
         it("should override service default_policy with method policy", () => {
             const svcOpts = createServiceOptions({ defaultPolicy: "deny" });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "OverridesPolicy", createMethodOptions({ policy: "allow" }));
+            const method = createFakeMethod(service, "OverridesPolicy", createMethodOptions({ policy: "allow" }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -198,7 +117,7 @@ describe("proto-reader", () => {
         it("should inherit service default_requires when method has no requires", () => {
             const svcOpts = createServiceOptions({ defaultRequires: { roles: ["user"] } });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "InheritsRequires");
+            const method = createFakeMethod(service, "InheritsRequires", undefined, { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -209,7 +128,7 @@ describe("proto-reader", () => {
         it("should override service default_requires with method requires", () => {
             const svcOpts = createServiceOptions({ defaultRequires: { roles: ["user"] } });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "OverridesRequires", createMethodOptions({ requires: { roles: ["admin"] } }));
+            const method = createFakeMethod(service, "OverridesRequires", createMethodOptions({ requires: { roles: ["admin"] } }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -220,7 +139,7 @@ describe("proto-reader", () => {
         it("should respect method public: false override on service public: true", () => {
             const svcOpts = createServiceOptions({ public: true });
             const service = createFakeService({ serviceOptions: svcOpts });
-            const method = createFakeMethod(service, "SecureMethod", createMethodOptions({ public: false }));
+            const method = createFakeMethod(service, "SecureMethod", createMethodOptions({ public: false }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -229,7 +148,7 @@ describe("proto-reader", () => {
 
         it("should cache resolved auth (same reference on second call)", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "CachedMethod", createMethodOptions({ public: true }));
+            const method = createFakeMethod(service, "CachedMethod", createMethodOptions({ public: true }), { register: true });
 
             const first = resolveMethodAuth(method);
             const second = resolveMethodAuth(method);
@@ -239,7 +158,7 @@ describe("proto-reader", () => {
 
         it("should ignore invalid policy strings", () => {
             const service = createFakeService();
-            const method = createFakeMethod(service, "InvalidPolicy", createMethodOptions({ policy: "invalid" }));
+            const method = createFakeMethod(service, "InvalidPolicy", createMethodOptions({ policy: "invalid" }), { register: true });
 
             const resolved = resolveMethodAuth(method);
 
@@ -250,9 +169,9 @@ describe("proto-reader", () => {
     describe("getPublicMethods()", () => {
         it("should return patterns for public methods", () => {
             const service = createFakeService({ typeName: "app.v1.AppService" });
-            createFakeMethod(service, "PublicMethod", createMethodOptions({ public: true }));
-            createFakeMethod(service, "PrivateMethod", createMethodOptions({ requires: { roles: ["admin"] } }));
-            createFakeMethod(service, "AnotherPublic", createMethodOptions({ public: true }));
+            createFakeMethod(service, "PublicMethod", createMethodOptions({ public: true }), { register: true });
+            createFakeMethod(service, "PrivateMethod", createMethodOptions({ requires: { roles: ["admin"] } }), { register: true });
+            createFakeMethod(service, "AnotherPublic", createMethodOptions({ public: true }), { register: true });
 
             const result = getPublicMethods([service]);
 
@@ -267,7 +186,7 @@ describe("proto-reader", () => {
 
         it("should return empty array when no public methods exist", () => {
             const service = createFakeService();
-            createFakeMethod(service, "PrivateMethod", createMethodOptions({ requires: { roles: ["admin"] } }));
+            createFakeMethod(service, "PrivateMethod", createMethodOptions({ requires: { roles: ["admin"] } }), { register: true });
 
             const result = getPublicMethods([service]);
 
@@ -277,8 +196,8 @@ describe("proto-reader", () => {
         it("should handle service-level public flag", () => {
             const svcOpts = createServiceOptions({ public: true });
             const service = createFakeService({ typeName: "pub.v1.PubService", serviceOptions: svcOpts });
-            createFakeMethod(service, "MethodA");
-            createFakeMethod(service, "MethodB");
+            createFakeMethod(service, "MethodA", undefined, { register: true });
+            createFakeMethod(service, "MethodB", undefined, { register: true });
 
             const result = getPublicMethods([service]);
 
@@ -287,12 +206,12 @@ describe("proto-reader", () => {
 
         it("should combine methods from multiple services", () => {
             const svc1 = createFakeService({ typeName: "svc1.v1.Svc1" });
-            createFakeMethod(svc1, "Public1", createMethodOptions({ public: true }));
-            createFakeMethod(svc1, "Private1");
+            createFakeMethod(svc1, "Public1", createMethodOptions({ public: true }), { register: true });
+            createFakeMethod(svc1, "Private1", undefined, { register: true });
 
             const svc2Opts = createServiceOptions({ public: true });
             const svc2 = createFakeService({ typeName: "svc2.v1.Svc2", serviceOptions: svc2Opts });
-            createFakeMethod(svc2, "AllPublic");
+            createFakeMethod(svc2, "AllPublic", undefined, { register: true });
 
             const result = getPublicMethods([svc1, svc2]);
